test: split machine plugins test
